requirements.txt
appnope==0.1.0
astroid==2.2.5
autopep8==1.4.3
backcall==0.1.0
cycler==0.10.0
decorator==4.4.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.17
jedi==0.13.3
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.3
mccabe==0.6.1
mysql-connector==2.2.9
numpy==1.16.2
pandas==0.24.2
parso==0.4.0
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.1
scipy==1.2.1
six==1.12.0
statsmodels==0.9.0
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.3.1
wcwidth==0.1.7
wrapt==1.11.1
